requirements.txt
numpy
pandas
seaborn
Pillow
scikit-learn
torch
torchvision
matplotlib
